Did shit....

+++ src/concepts.hpp
@@ -1,11 +1,13 @@
 #ifndef LOOPR_CONCEPTS
 #define LOOPR_CONCEPTS
 
+#include <concepts>
+
 ///<summary> 
 /// Any type that can be multiplied, divided by, added to,  subtracted from a float
 ///</summary>
-template <typename T> 
-concept Arithmetic = requires(T a, float b) {
+template <typename T>
+concept Arithmetic = requires(T a, double b) {
     a * b;
     a / b;
     a + b;
